third_party/WebKit/LayoutTests/platform/win/http/tests/security/suborigins/crossorigin/suborigin-cors-fetch-expected.txt

CONSOLE ERROR: line 32: Fetch API cannot load http://127.0.0.1:8000/security/resources/cors-script.php?value=ran&cors=*. Request mode is "same-origin" but the URL's origin is not same as the request origin http-so://foobar.127.0.0.1:8000.
CONSOLE ERROR: Fetch API cannot load http://127.0.0.1:8000/security/resources/cors-script.php?value=ran. The 'Access-Control-Allow-Origin' header has a value 'http://127.0.0.1:8000' that is not equal to the supplied origin. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access. Have the server send the header with a valid value, or, if an opaque response serves your needs, set the request's mode to 'no-cors' to fetch the resource with CORS disabled.
CONSOLE ERROR: Fetch API cannot load http://127.0.0.1:8000/security/resources/cors-script.php?credentials=true&value=ran. The 'Access-Control-Allow-Origin' header has a value 'http://127.0.0.1:8000' that is not equal to the supplied origin. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access. Have the server send the header with a valid value, or, if an opaque response serves your needs, set the request's mode to 'no-cors' to fetch the resource with CORS disabled.
CONSOLE ERROR: Fetch API cannot load http://127.0.0.1:8000/security/resources/cors-script.php?value=ran&cors=false. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access. If an opaque response serves your needs, set the request's mode to 'no-cors' to fetch the resource with CORS disabled.
This is a testharness.js-based test.
PASS Fetch: anonymous, ACAO: http://127.0.0.1:8000 
PASS Fetch: anonymous, ACAO: * 
PASS Fetch: use-credentials, ACAO: http://127.0.0.1:8000 
PASS Fetch: anonymous, CORS-ineligible resource 
PASS Fetch: same-origin, ACAO: *  
Harness: the test ran to completion.
